requirements.txt
openai
openai[embeddings]
python-dotenv
pandas
tiktoken
jupyter
scikit-learn
matplotlib
numpy
plotly